Extract depth to a separate module
