+++ flatgfa/src/cmds.rs
@@ -1,10 +1,11 @@
 use crate::flatgfa::{self, Handle, Link, Orientation, Path, Segment};
 use crate::memfile;
+use crate::ops;
 use crate::pool::{self, Id, Span, Store};
 use crate::{GFAStore, HeapFamily};
 use argh::FromArgs;
 use rayon::iter::ParallelIterator;
-use std::collections::{HashMap, HashSet};
+use std::collections::HashMap;
 
 /// print the FlatGFA table of contents
 #[derive(FromArgs, PartialEq, Debug)]
@@ -410,22 +411,8 @@ impl<'a> SubgraphBuilder<'a> {
 pub struct Depth {}
 
 pub fn depth(gfa: &flatgfa::FlatGFA) {
-    // Initialize node depth
-    let mut depths = vec![0; gfa.segs.len()];
-    // Initialize uniq_paths
-    let mut uniq_paths = Vec::<HashSet<usize>>::new();
-    uniq_paths.resize(gfa.segs.len(), HashSet::new());
-    // do not assume that each handle in `gfa.steps()` is unique
-    for (idx, path) in gfa.paths.all().iter().enumerate() {
-        for step in &gfa.steps[path.steps] {
-            let seg_id = step.segment().index();
-            // Increment depths
-            depths[seg_id] += 1;
-            // Update uniq_paths
-            uniq_paths[seg_id].insert(idx);
-        }
-    }
-    // print out depth and depth.uniq
+    let (depths, uniq_paths) = ops::depth::depth(gfa);
+
     println!("#node.id\tdepth\tdepth.uniq");
     for (id, seg) in gfa.segs.items() {
         let name: u32 = seg.name as u32;
