+++ flatgfa/src/lib.rs
@@ -5,6 +5,7 @@ pub mod gaf;
 pub mod gfaline;
 pub mod memfile;
 pub mod namemap;
+pub mod ops;
 pub mod parse;
 pub mod pool;
 pub mod print;
